"check_style_score -> 0 of 5 passed. style score: 0/1 "

+++ CSPP-2 assignments/M2/Assignment-3/MySolution.java
@@ -19,11 +19,11 @@ public final class MySolution {
     *@param args is parameter for this function.
     **/
     public static void main(final String[] args) {
-        double a, b;
-        double pow = 1;
+        float a, b;
+        float pow = 1;
         Scanner s = new Scanner(System.in);
-        a = s.nextDouble();
-        b = s.nextDouble();
+        a = s.nextFloat();
+        b = s.nextFloat();
         pow = power(a, b);
         System.out.println(pow);
     }
@@ -33,7 +33,7 @@ public final class MySolution {
     *@param b is parameter for this function.
     *@return int
     **/
-    public static double power(final double a, final double b) {
+    public static float power(final float a, final float b) {
         if (b != 0) {
             return (a * power(a, b - 1));
         } else {
